% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynx.R
\docType{data}
\name{lynx}
\alias{lynx}
\title{Canadian Lynx Data.}
\format{
A data list with 114 rows containing the count of Canadian Lynx from 1821-1934.
}
\source{
https://www.encyclopediaofmath.org /index.php/Canadian_lynx_data
}
\usage{
data(lynx)
}
\description{
Annual record of the number of the Canadian
Lynx `trapped' in the Mackenzie River
district of the North-West Canada for the period 1821-1934.
}
\keyword{datasets}
